refactor(client): :recycle: Extracted new LocationMenuItem component
