+++ client/src/components/Pages/recherche/LocationDropdown/LocationMenuItem/LocationMenuItem.module.css
@@ -1,0 +1,74 @@
+.container {
+  /* Layout */
+  display: flex;
+  width: 320px;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 8px;
+}
+
+.selectedLocationItem {
+  /* Layout */
+  display: flex;
+  padding: 8px 8px 8px 10px;
+  align-items: flex-start;
+  gap: 8px;
+  align-self: stretch;
+}
+
+.locationItem {
+  /* Layout */
+  display: flex;
+  padding: 8px 10px;
+  justify-content: space-between;
+  align-items: center;
+  align-self: stretch;
+
+  /* Style */
+  background: var(--background-alt-blue-france);
+}
+
+.locationIcon {
+  /* Layout */
+  display: flex;
+  padding: 4px;
+  align-items: center;
+  gap: 10px;
+
+  width: 24px;
+  height: 24px;
+  padding: 4px;
+  display: inline-block; /* Ensure the element respects the width and height */
+
+  &::before {
+    content: "";
+    display: block; /* Ensure the pseudo-element respects the width and height */
+    width: 100%;
+    height: 100%;
+    -webkit-mask-size: 16px 16px; /* Ensure the mask image is sized correctly */
+    mask-size: 16px 16px; /* For non-webkit browsers */
+  }
+}
+
+.button {
+  /* Override the default button styles */
+  padding-block: 0px;
+  padding-inline: 0px;
+
+  /* Layout */
+  display: flex;
+  align-items: flex-start;
+  gap: 8px;
+  color: var(--background-action-high-blue-france-hover);
+}
+
+.buttonText {
+  color: var(--background-action-high-blue-france-hover);
+
+  /* 2.Corps de texte/M - Texte standard/Desktop & Mobile - Medium */
+  font-family: Marianne;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 24px; /* 150% */
+}
