+++ client/src/components/Pages/recherche/LocationDropdown/LocationMenuItem/index.ts
@@ -1,0 +1,3 @@
+import LocationMenuItem from "./LocationMenuItem";
+
+export default LocationMenuItem;
